# backend/src/test/resources/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 프로젝트 생성 공통기능 API
  version: 1.0.0
  description: 한국교원대학교 교수업적평가시스템 공통기능 1차 범위 HTTP API 계약
servers:
  - url: /
    description: 상대경로 /api 사용 계약
paths:
...
  /api/admin/menu-permissions:
    get:
      operationId: listMenuPermissions
      summary: 메뉴 권한 조회
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 검증 오류
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
          description: 페이지 번호
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
          description: 페이지 크기
        - name: targetType
          in: query
          required: false
          schema:
            type: string
          description: targetType
        - name: targetId
          in: query
          required: false
          schema:
            type: string
          description: targetId
        - name: filter
          in: query
          required: false
          schema:
            type: string
      x-related-requirements:
        - REQ-038
        - REQ-039
        - REQ-040
        - REQ-041
        - REQ-042
        - REQ-043
        - REQ-004
        - REQ-006
        - REQ-007
        - REQ-068
        - REQ-070
        - REQ-076
        - REQ-077
        - REQ-083
      security:
        - SessionCookie: []
      x-roles:
        - R09
    put:
      operationId: saveMenuPermissions
      summary: 메뉴 접근권한 저장
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 검증 오류
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuPermissionSaveRequest'
      x-related-requirements:
        - REQ-036
        - REQ-037
        - REQ-041
        - REQ-042
        - REQ-043
        - REQ-063
        - REQ-003
        - REQ-004
        - REQ-006
        - REQ-007
        - REQ-008
        - REQ-065
        - REQ-073
        - REQ-070
        - REQ-076
        - REQ-077
        - REQ-083
      security:
        - SessionCookie: []
      x-roles:
        - R09
      x-business-rules:
        - 메뉴 접근 권한 설정 요구의 입력 필드는 서버에서 검증한다
        - 인증된 R09 관리자만 메뉴 접근권한 저장 요청을 수행할 수 있다
      x-side-effects:
        - 메뉴 접근권한 저장 결과를 관련 table에 저장하고 변경 메타정보를 기록한다
      x-required-tests:
        - 'happy: 유효한 메뉴 접근권한 저장 요청 -> 후속 조회에서 REQ-036 관련 table 변경값이 확인된다'
        - 'auth: 인증 없는 메뉴 접근권한 저장 요청 -> 401 ApiError가 반환된다'
        - 'validation: 필수 field 누락 메뉴 접근권한 저장 요청 -> 해당 field-level ApiError가 반환된다'
        - 'business: 메뉴 접근권한 저장 업무 규칙 위반 -> 기존 상태가 변경되지 않는다'
        - 'side-effect: 메뉴 접근권한 저장 성공 -> 변경 처리자와 처리일시가 관련 row에 남는다'
...
  /api/admin/code-groups:
    get:
      operationId: listCodeGroups
      summary: 코드그룹 조회
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 검증 오류
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
          description: 페이지 번호
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
          description: 페이지 크기
        - name: groupIdFilter
          in: query
          required: false
          schema:
            type: string
          description: groupIdFilter
        - name: filter
          in: query
          required: false
          schema:
            type: string
      x-related-requirements:
        - REQ-054
        - REQ-055
        - REQ-068
        - REQ-070
        - REQ-076
        - REQ-077
        - REQ-086
      security:
        - SessionCookie: []
      x-roles:
        - R09
    post:
      operationId: createCodeGroup
      summary: 코드그룹 등록
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 검증 오류
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeGroupRequest'
      x-related-requirements:
        - REQ-053
        - REQ-056
        - REQ-003
        - REQ-008
        - REQ-065
        - REQ-073
        - REQ-070
        - REQ-076
        - REQ-077
        - REQ-086
      security:
        - SessionCookie: []
      x-roles:
        - R09
      x-business-rules:
        - 코드그룹 등록·수정 요구의 입력 필드는 서버에서 검증한다
        - 인증된 R09 관리자만 코드그룹 등록 요청을 수행할 수 있다
      x-side-effects:
        - 코드그룹 등록 결과를 관련 table에 저장하고 변경 메타정보를 기록한다
      x-required-tests:
        - 'happy: 유효한 코드그룹 등록 요청 -> 후속 조회에서 REQ-053 관련 table 변경값이 확인된다'
        - 'auth: 인증 없는 코드그룹 등록 요청 -> 401 ApiError가 반환된다'
        - 'validation: 필수 field 누락 코드그룹 등록 요청 -> 해당 field-level ApiError가 반환된다'
        - 'business: 코드그룹 등록 업무 규칙 위반 -> 기존 상태가 변경되지 않는다'
        - 'side-effect: 코드그룹 등록 성공 -> 변경 처리자와 처리일시가 관련 row에 남는다'
  /api/admin/code-groups/{groupId}:
    put:
      operationId: updateCodeGroup
      summary: 코드그룹 수정
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 요청 검증 오류
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '401':
          description: 인증 필요
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
        '403':
          description: 권한 없음
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      error:
                        $ref: '#/components/schemas/ApiError'
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
          description: groupId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeGroupRequest'
      x-related-requirements:
        - REQ-053
        - REQ-056
        - REQ-003
        - REQ-008
        - REQ-065
        - REQ-073
        - REQ-074
        - REQ-070
        - REQ-076
        - REQ-077
        - REQ-086
      security:
        - SessionCookie: []
      x-roles:
        - R09
      x-business-rules:
        - 코드그룹 등록·수정 요구의 입력 필드는 서버에서 검증한다
        - 인증된 R09 관리자만 코드그룹 수정 요청을 수행할 수 있다
      x-side-effects:
        - 코드그룹 수정 결과를 관련 table에 저장하고 변경 메타정보를 기록한다
      x-required-tests:
        - 'happy: 유효한 코드그룹 수정 요청 -> 후속 조회에서 REQ-053 관련 table 변경값이 확인된다'
        - 'auth: 인증 없는 코드그룹 수정 요청 -> 401 ApiError가 반환된다'
        - 'validation: 필수 field 누락 코드그룹 수정 요청 -> 해당 field-level ApiError가 반환된다'
        - 'business: 코드그룹 수정 업무 규칙 위반 -> 기존 상태가 변경되지 않는다'
        - 'side-effect: 코드그룹 수정 성공 -> 변경 처리자와 처리일시가 관련 row에 남는다'
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: HttpOnly SameSite=Lax session cookie
  schemas:
    ApiResponse:
      type: object
      required:
        - success
        - meta
      properties:
        success:
          type: boolean
        data: {}
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          type: object
          additionalProperties: true
    ApiError:
      type: object
      required:
        - code
        - message
        - fields
      properties:
        code:
          type: string
        message:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
...
    MenuPermissionSaveRequest:
      type: object
      required:
        - targetType
        - targetId
        - menuId
        - accessAllowed
      properties:
        targetType:
          $ref: '#/components/schemas/TargetTypeEnum'
        targetId:
          type: string
        menuId:
          type: string
        accessAllowed:
          $ref: '#/components/schemas/AccessAllowedEnum'
        changeReason:
          type: string
...
    CodeGroupRequest:
      type: object
      required:
        - groupId
        - groupName
        - managingDepartment
      properties:
        groupId:
          type: string
        groupName:
          type: string
        description:
          type: string
        managingDepartment:
          type: string
        systemUseYn:
          $ref: '#/components/schemas/SystemUseYnEnum'
        changeReason:
          type: string
...
    SystemUseYnEnum:
      type: string
      enum:
        - Y
        - N
...
    TargetTypeEnum:
      type: string
      enum:
        - ROLE
        - ORGANIZATION
        - USER
    AccessAllowedEnum:
      type: string
      enum:
        - ALLOW
        - DENY
...